+++ src/n8n/nodes/30-calendar-create/prepare-finalize-request__4c9a5cd8-7c13-4ad8-8d1c-a10f2f23520b.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('./prepare-finalize-request__e114ec84-f1a4-4796-84a0-d4ec1229b304.js');