+++ src/n8n/nodes/32-calendar-report/build-report-window__1d7fa7c9-3ac6-4b7e-bf0a-6e2e7789f31a.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('../32-calendar-daily-report/build-report-window__ecf30966-6990-4e1d-9611-729f02c53a61.js');