+++ src/n8n/nodes/32-calendar-report/format-calendar-report-message__58f6c53c-5dad-4d29-93d0-00dc8f7d5683.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('../32-calendar-daily-report/format-calendar-report-message__dd8df3c8-4be6-442f-b28c-05fa47e25751.js');